requirements.txt
# requirements.txt

Flask==2.2.2
python-dotenv==0.21.0
psycopg2
Flask-Cors==3.0.10
PyJWT==2.6.0
bcrypt==4.0.1
requests==2.28.1